Consolidate two pages back onto one page
